refactor(types): Move Name and TreeState to types.tsx
