+++ treeAct-view/src/types.tsx
@@ -1,4 +1,9 @@
-export interface TreeNode {
+import { NewTreeNode } from "../../common/types";
+
+export type Name = {
+  id: number;
   name: string;
-  children?: TreeNode[];
-}
+};
+
+// Type of whole tree data
+export type TreeState = { treeData: NewTreeNode[] };
